requirements.txt
python-telegram-bot==13.15
requests==2.28.2
Flask==2.2.5